+++ api/models/device.py
@@ -43,9 +43,9 @@ class Device(models.Model):
         get_latest_by = '-last_update'
 
     @staticmethod
-    def from_json(spark, devices_json):
+    def from_json_list(spark, devices_json):
 
-        logger.debug("Convert json devices into Device object")
+        logger.debug("Convert json devices into Device objects")
         try:
             logger.debug("Received: " + devices_json)
             devices_list = json.loads(devices_json)
@@ -64,7 +64,7 @@ class Device(models.Model):
 
                     device.spark = spark
                     device.type = device_dic.get('type')
-                    device.value = device_dic.get('value')
+                    device.value = device_dic.get('value').lstrip()
                     device.is_invert = device_dic_hardware.get('is_invert')
                     device.offset = device_dic_hardware.get('offset')
                     device.is_deactivate = device_dic_hardware.get('is_deactivate')
@@ -74,7 +74,7 @@ class Device(models.Model):
                     logger.debug(device.__str__())
 
                 except ObjectDoesNotExist:
-                    device = Device.create(spark, device_dic.get('type'), device_dic.get('value'),
+                    device = Device.create(spark, device_dic.get('type'), device_dic.get('value').lstrip(),
                                            device_dic_hardware.get('pin_nr'),
                                            device_dic_hardware.get('hw_address'), device_dic_hardware.get('offset'),
                                            device_dic_hardware.get('is_invert'),
@@ -89,6 +89,29 @@ class Device(models.Model):
         Device.objects.filter(spark=spark, last_update__lt=(timezone.now() - timedelta(minutes=1))).delete()
 
         return Device.objects.filter(spark=spark)
+
+    @staticmethod
+    def from_json(device, device_json):
+
+        logger.debug("Convert json devices into Device objects")
+        try:
+            logger.debug("Received: " + device_json)
+            device_dic = json.loads(device_json)
+        except ValueError:
+            return None
+
+        device_dic_hardware = device_dic.get('hardware')
+
+        device.spark = device.spark
+        device.type = device_dic.get('type')
+        device.value = device_dic.get('value').lstrip()
+        device.is_invert = device_dic_hardware.get('is_invert')
+        device.offset = device_dic_hardware.get('offset')
+        device.is_deactivate = device_dic_hardware.get('is_deactivate')
+        device.last_update = timezone.now()
+        device.save()
+
+        logger.debug(device.__str__())
 
     @classmethod
     def create(cls, spark, hw_type, value, pin_nr, hw_address, offset, is_invert, is_deactivate):