+++ CHANGELOG.md
@@ -14,3 +14,4 @@ v0.1
 - Receive the Spark date time setting and store in the database
 - Send the web server date time as part of the Spark Info message
 - Toggle Device Actuator on a specific Spark provided in URL and return new state
+- Request Sensor/Actuator status from a specific Spark provided URL